pyproject.toml
[tool.poetry]
name = "ndsh"
version = "0.1.0"
description = "Remote shell over NDN"
authors = ["Xinyu Ma <ma.xinyu.26a@kyoto-u.jp>"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-ndn = {version="^0.3-3", extras=["pyca"]}
# python-ndn = {path="../python-ndn/", extras=["pyca"], develop = true}
shellous = "^0.25.0"
aiohttp = "^3.8.4"
aiohttp-jinja2 = "^1.5.1"
python-socketio = "^5.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
